Add -s/--silent option
I have noticed a problem where the script stops executing in the
background once a newline is written to stderr, when run in neovim 0.7.0
and python installed via Homebrew (on OSX).  When confirming this in vim
I also noticed the stderr stuff was being printed on top of the vim UI;
running it silently in vim is a way to fix or sidestep both of these
problems.